cli: Add rootless logic to container mappings
Remove the ctrsMapTreePath global variable, and make a Getter
function that checks whether a user is running the runtime as
root or not, and if root returns the default value, otherwise
returns the rootless path. Modify functions using ctrsMapTreePath
to pass it in as a variable. This will allow test cases to use
a custom ctrsMapTreePath.

Signed-off-by: Gabi Beyer <gabrielle.n.beyer@intel.com>